requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
psycopg2-binary>=2.9.5
pydantic>=2.0.0
pytest-cov>=6.0.0
httpx>=0.28.0

# Optional dependencies for development
# uvicorn[standard]  # Includes websockets, httptools, watchfiles for improved performance 